Devolve the automatic running down to workflow
Composite will build its innards only once
